BC-1951 Add video and audio browser command
