+++ static/scripts/ckeditor/plugins/ckeditor5-filebrowser/src/file-browser-modal.js
@@ -1,6 +1,4 @@
-import Command from '@ckeditor/ckeditor5-core/src/command';
-
-const createFilebrowserModal = (editor, t, dialogTitle, onCreate, additionalInput = '') => {
+export default function createFilebrowserModal(editor, t, dialogTitle, onCreate, additionalInput = '') {
 	const ckeditorFilebrowserDialog = $('.ckeditor-filebrowser-dialog');
 	populateModalForm(ckeditorFilebrowserDialog, {
 		title: dialogTitle,
@@ -29,28 +27,4 @@ const createFilebrowserModal = (editor, t, dialogTitle, onCreate, additionalInpu
 		const dialogPageUrl = `${editor.config.get('filebrowser.browseUrl')}?CKEditor=true`;
 		window.open(dialogPageUrl, '_blank', 'width=700, height=500');
 	});
-};
-
-export default class ImageBrowserCommand extends Command {
-	execute() {
-		const additionalInput = `<br><label for="alt-text-input">${this.editor.t('Alternative Text')}:</label>
-		  <input type="text" id="alt-text-input">`;
-		const dialogTitle = this.editor.t('Image Properties');
-		const onCreate = () => {
-			const imageUrl = document.getElementById('url-input').value;
-			if (!imageUrl) return;
-			const imageAltText = document.getElementById('alt-text-input').value;
-			this.editor.model.change((writer) => {
-				const imageElement = writer.createElement('image', {
-					src: imageUrl,
-					alt: imageAltText,
-				});
-				const lastOpenedEditorId = document.getElementById('editor-id').value;
-				if (lastOpenedEditorId === this.editor.id) {
-					this.editor.model.insertContent(imageElement, this.editor.model.document.selection);
-				}
-			});
-		};
-		createFilebrowserModal(this.editor, this.editor.t, dialogTitle, onCreate, additionalInput);
-	}
 }
