+++ static/scripts/topicEdit.js
@@ -458,6 +458,8 @@ class TopicText extends TopicBlock {
 
 		if (topicId === undefined) {
 			editor.commands.get('imagebrowser').forceDisabled();
+			editor.commands.get('audiobrowser').forceDisabled();
+			editor.commands.get('videobrowser').forceDisabled();
 		}
 	}
